Desktop/novio/requirements.txt
# Core
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1

# LangGraph & LangChain
langgraph>=1.0.0
langchain-core>=1.0.0
langchain-ollama>=1.0.0
langchain-community>=0.4.0
langchain-huggingface>=1.0.0

# Vector Store & Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=5.0.0

# PDF Processing
PyPDF2==3.0.1
